[MOV] sale_timesheet: move project, product.template fields to sale_project
The `product.template` service policies that do not involve timesheets should be available with `sale_project` and without `sale_timesheet`.
This commit adds the `service_policy` field and moves those policies to `sale_project`.

It also moves the `allow_billable` field on project.project to `sale_project`.

Task-2558889

Closes https://github.com/odoo/odoo/pull/81684
