+++ addons/sale_project/models/product.py
@@ -5,6 +5,19 @@ from odoo import api, fields, models, _
 from odoo.exceptions import ValidationError
 
 
+SERVICE_POLICY = [
+    # (service_policy, string)
+    ('ordered_timesheet', 'Prepaid/Fixed Price'),
+    ('delivered_manual', 'Based on Delivered Quantity (Manual)'),
+]
+SERVICE_TO_GENERAL = {
+    # service_policy: (invoice_policy, service_type)
+    'ordered_timesheet': ('order', 'manual'),
+    'delivered_manual': ('delivery', 'manual'),
+}
+GENERAL_TO_SERVICE = {value: key for key, value in SERVICE_TO_GENERAL.items()}
+
+
 class ProductTemplate(models.Model):
     _inherit = 'product.template'
 
@@ -28,31 +41,70 @@ class ProductTemplate(models.Model):
         'project.project', 'Project Template', company_dependent=True, copy=True,
         domain="[('company_id', '=', current_company_id)]",
         help='Select a billable project to be the skeleton of the new created project when selling the current product. Its stages and tasks will be duplicated.')
+    service_policy = fields.Selection(SERVICE_POLICY, string="Service Invoicing Policy", compute='_compute_service_policy', inverse='_inverse_service_policy')
 
     @api.depends('service_tracking', 'type')
     def _compute_product_tooltip(self):
         super()._compute_product_tooltip()
         for record in self.filtered(lambda record: record.type == 'service'):
-            if record.service_tracking == 'no':
-                record.product_tooltip = _(
-                    "Invoice ordered quantities as soon as this service is sold."
-                )
-            elif record.service_tracking == 'task_global_project':
-                record.product_tooltip = _(
-                    "Invoice as soon as this service is sold, and create a task in an existing "
-                    "project to track the time spent."
-                )
-            elif record.service_tracking == 'task_in_project':
-                record.product_tooltip = _(
-                    "Invoice ordered quantities as soon as this service is sold, and create a "
-                    "project for the order with a task for each sales order line to track the time"
-                    " spent."
-                )
-            elif record.service_tracking == 'project_only':
-                record.product_tooltip = _(
-                    "Invoice ordered quantities as soon as this service is sold, and create an "
-                    "empty project for the order to track the time spent."
-                )
+            if record.service_policy == 'ordered_timesheet':
+                if record.service_tracking == 'no':
+                    record.product_tooltip = _(
+                        "Invoice ordered quantities as soon as this service is sold."
+                    )
+                elif record.service_tracking == 'task_global_project':
+                    record.product_tooltip = _(
+                        "Invoice as soon as this service is sold, and create a task in an existing "
+                        "project to track the time spent."
+                    )
+                elif record.service_tracking == 'task_in_project':
+                    record.product_tooltip = _(
+                        "Invoice ordered quantities as soon as this service is sold, and create a "
+                        "project for the order with a task for each sales order line to track the time"
+                        " spent."
+                    )
+                elif record.service_tracking == 'project_only':
+                    record.product_tooltip = _(
+                        "Invoice ordered quantities as soon as this service is sold, and create an "
+                        "empty project for the order to track the time spent."
+                    )
+            elif record.service_policy == 'delivered_manual':
+                if record.service_tracking == 'no':
+                    record.product_tooltip = _(
+                        "Sales order lines define milestones of the project to invoice by setting "
+                        "the delivered quantity."
+                    )
+                elif record.service_tracking == 'task_global_project':
+                    record.product_tooltip = _(
+                        "Sales order lines define milestones of the project to invoice by setting "
+                        "the delivered quantity. Create a task in an existing project to track the"
+                        " time spent."
+                    )
+                elif record.service_tracking == 'task_in_project':
+                    record.product_tooltip = _(
+                        "Sales order lines define milestones of the project to invoice by setting "
+                        "the delivered quantity. Create an empty project for the order to track "
+                        "the time spent."
+                    )
+                elif record.service_tracking == 'project_only':
+                    record.product_tooltip = _(
+                        "Sales order lines define milestones of the project to invoice by setting "
+                        "the delivered quantity. Create a project for the order with a task for "
+                        "each sales order line to track the time spent."
+                    )
+
+    @api.depends('invoice_policy', 'service_type', 'type')
+    def _compute_service_policy(self):
+        for product in self:
+            product.service_policy = GENERAL_TO_SERVICE.get((product.invoice_policy, product.service_type), False)
+            if not product.service_policy and product.type == 'service':
+                product.service_policy = 'ordered_timesheet'
+
+    @api.onchange('service_policy')
+    def _inverse_service_policy(self):
+        for product in self:
+            if product.service_policy:
+                product.invoice_policy, product.service_type = SERVICE_TO_GENERAL.get(product.service_policy, (False, False))
 
     @api.constrains('project_id', 'project_template_id')
     def _check_project_and_template(self):
