+++ addons/sale_project/models/project.py
@@ -10,6 +10,7 @@ from odoo.exceptions import ValidationError, AccessError
 class Project(models.Model):
     _inherit = 'project.project'
 
+    allow_billable = fields.Boolean("Billable", help="Invoice your time and material from tasks.")
     sale_line_id = fields.Many2one(
         'sale.order.line', 'Sales Order Item', copy=False,
         compute="_compute_sale_line_id", store=True, readonly=False, index='btree_not_null',
