+++ addons/sale_project/views/product_views.xml
@@ -6,7 +6,12 @@
         <field name="model">product.template</field>
         <field name="inherit_id" ref="sale.product_template_form_view"/>
         <field name="arch" type="xml">
+            <field name="invoice_policy" position="attributes">
+                <attribute name="invisible">False</attribute>
+                <attribute name="attrs">{'invisible': [('type', '=', 'service')]}</attribute>
+            </field>
             <field name="invoice_policy" position="after">
+                <field name="service_policy" string="Invoicing Policy" attrs="{'invisible': [('type', '!=', 'service')], 'required': [('type', '=', 'service')]}"/>
                 <field name="service_tracking" required="1" attrs="{'invisible': [('type', '!=', 'service')]}"/>
             </field>
             <field name="product_tooltip" position="after">
